refactor:pageSize logic extracted to custom hook
